Configure packages in chroot in rootfs.
As from kernel version 4.14 the v3 capabilities xattr landed it is
no longer needed to patch setcap.

By using chroot to configure packages patching of the xbps-triggers is
no longer needed. That also makes patching of void-packges and building
of several packages also unnecessary.

Builder tool is replaced by unshare-chroot (which probably should be renamed)
it is used for masking chgrp and chown binaries as builder before,
but also to run a chroot under a unshare with user mapped as root.
Void's xbps-uunshare seems not enough to do this.